% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt2irtStan-package.R
\docType{package}
\name{mpt2irtStan-package}
\alias{mpt2irtStan-package}
\alias{mpt2irtStan}
\title{The 'mpt2irtStan' package.}
\description{
This package provides the Stan code for the mpt2irt package and is most likely only useful in combination with mpt2irt.
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.18.2. https://mc-stan.org

Plieninger, H., & Heck, D. W. (2018). A new model for
  acquiescence at the interface of psychometrics and cognitive psychology.
  Multivariate Behavioral Research, 53(3), 633-654.
  doi:10.1080/00273171.2018.1469966
}
